requirements.txt
pydantic>=2.6
pdf2image>=1.17
pillow>=10
pytesseract>=0.3.10
python-magic